refactor: created informations component
