+++ src/pages/Profile/styles.ts
@@ -1,46 +1,22 @@
-import { MapPin } from "react-feather";
 import styled from "styled-components";
+import { Screen as GlobalScreen, Text as GlobalText } from "../../styles/global";
 
-const ProfileHeader = styled.div`
-  display: flex;
-  width: 100%;
-  gap: 24px;
-  align-items: center;
-
-  @media (max-width: ${(props) => props.theme.breakpoints.TABLET}) {
-    flex-direction: column;
-    gap: 8px;
-  }
-`;
-
-const HeaderText = styled.div`
-  display: flex;
+const Screen = styled(GlobalScreen)`
+  justify-content: flex-start;
   flex-direction: column;
-  gap: 4px;
-  justify-content: center;
-
-  @media (max-width: ${(props) => props.theme.breakpoints.TABLET}) {
-    align-items: center;
-  }
 `;
 
-const DescriptionArea = styled.div`
+const FeedHeader = styled.div`
   display: flex;
-  width: 100%;
-  flex-direction: column;
-  gap: 16px;
-`;
-
-const LocationArea = styled.div`
-  display: flex;
-  gap: 4px;
+  width: 480px;
+  justify-content: space-between;
+  gap: 8px;
   align-items: center;
 `;
 
-const PinIcon = styled(MapPin)`
-  color: ${(props) => props.theme.colors.GREY};
-  width: 18px;
-  height: 18px;
+const Text = styled(GlobalText)`
+  font-weight: bold;
+  white-space: nowrap;
 `;
 
-export { ProfileHeader, HeaderText, DescriptionArea, LocationArea, PinIcon };
+export { Screen, FeedHeader, Text };
